git-svn-id: https://grub4dos-chenall.googlecode.com/svn/branches/yaya-0.4.6a@291 ac2b6734-dafe-11de-aa1f-bbd297b2055c

+++ grub/asmstub.c
@@ -955,7 +955,7 @@ hex_dump (void *buf, size_t size)
 
 int
 biosdisk (int subfunc, int drive, struct geometry *geometry,
-	  int sector, int nsec, int segment)
+	  unsigned long long sector, unsigned long nsec, int segment)
 {
   char *buf;
   int fd = geometry->flags;
